Zswap Zpool allocator: use alternative if allocator is not available
Firstly, z3fold has been deprecated since kernel 6.12. So it can be
expected that some actual systemd-swap users will hit a wall in the near
future either because the z3fold module will not be built by default or
because it will have been stripped out of the kernel entirely.

Secondly, talks are ongoing about doing the same for zbud.

Finally, the user should be both warned about missing modules preventing
swapFC from working as expected, but be supported in maintaining the
functionality available by using alternative allocators.

Thus, a check is applied to validate if the requested allocator is
available. If it is not, systemd-swap should either use the allocator
already set or default to zsmalloc (which is expected to be available if
Zswap was built in the kernel).

Signed-off-by: Alexandre Demers <alexandre.f.demers@gmail.com>